DX12: fix many black block on screen

+++ DX12/D3D12Research/D3D12/ClusteredForward.cpp
@@ -42,7 +42,7 @@ void ClusteredForward::OnSwapchainCreated(int windowWidth, int windowHeight)
 
 	float nearZ = 2.0f;
 	float farZ = 500.0f;
-	Matrix projection = XMMatrixPerspectiveFovLH(XM_PIDIV4, windowWidth / windowHeight, nearZ, farZ);
+	Matrix projection = XMMatrixPerspectiveFovLH(XM_PIDIV4, (float)windowWidth / windowHeight, nearZ, farZ);
 
 	// create AABBs
 	{
